searchbar firing
